JustSplit/requirements.txt
aiosmtplib==2.0.2
alembic==1.11.1
anyio==3.7.0
asyncpg==0.27.0
bcrypt==4.0.1
blinker==1.6.2
boto3==1.27.0
botocore==1.30.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
decouple==0.0.7
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.2
fastapi==0.99.1
fastapi-jwt-auth==0.5.0
fastapi-mail==1.3.1
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.1
passlib==1.7.4
Pillow==10.0.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pydantic==1.10.11
PyJWT==1.7.1
pyscopg2==66.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
razorpay==1.3.0
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
twilio==6.63.2
typing_extensions==4.7.1
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
